requirements.txt
Django==5.2.3
pandas==2.3.0
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2025.2
sqlparse==0.5.3
numpy==2.3.1